+++ app/models/vacation.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+class Vacation < ApplicationRecord
+  belongs_to :user
+end