+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Added
+- excluding vacation time from calculating of average time
+- editing vacation time for users
+
 ## [1.2.3] - 2023-02-26
 ### Fixed
 - awarding for comment creating